[Refactor]#2 NetworkUtil 패키지 이동
- util to network
